Update patch set 4
Patch Set 4:

I suppose that if ssh works reliably (and the dance to start it up and get each side to agree on a key isn’t too bad), then that’d be even better than this extra custom code, neat as it is. Can you share a link to what’s happening in Chrome for this?

Patch-set: 4
Reviewer: Mark Mentovai <1174099@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Work-in-progress: false