Update patch set 4
Patch Set 4:

> Patch Set 4:
> 
> I suppose that if ssh works reliably (and the dance to start it up and get each side to agree on a key isn’t too bad), then that’d be even better than this extra custom code, neat as it is. Can you share a link to what’s happening in Chrome for this?

Much of it is here: https://chromium-review.googlesource.com/c/chromium/src/+/882441

It is a bit more involved than I'd like to set up the keys though. There's a default dev ssh config that the in-tree Fuchsia builds set up, but James felt it was a bad idea to include that ssh config in the SDK as it'd leave anything built with the SDK wide open, which makes sense.

It'd be nice to have some sort of one step helper dev-only key generation script in Fuchsia... but that doesn't exist at the moment.

Patch-set: 4