Update patch set 7
Patch Set 7:

Geez, this is a mess. Now the debug simulated one doesn't work because it's 2 bytes farther than the expected range. I investigated why, and the release ones are compact enough that they work as expected, and using WriteFile() instead of CheckedWriteFile() avoids the inlining so the final result is smaller.

In debug though, the combination of the macro expansion of CRASHPAD_SIMULATE_CRASH along with the other code is slightly too long.

Why did it work before? Well, there didn't inline any of CheckedWriteFile(), and there are fewer arguments to CheckedWriteFile() than ::WriteFile(), so slightly less code between CaptureContext and CRASHPAD_SIMULATE_CRASH. Indeed, it worked on Win10 when I was passing null for lpNumberOfBytesWritten because that extra address lea/push puts it over the edge.

Anyway, I think it's reasonable to increase the threshold a little and add even more rationalization to the already-long comment.

Patch-set: 7
Reviewer: Scott Graham <1144260@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Commit-Queue=0