fuchsia: Prototype getting g_crashpad_info address through via .note
Embeds the address of g_crashpad_info into a .note section (which is
readable by the generic code to read notes in ElfImageReader).
Unfortunately because the note section is in libclient.a, it would
normally be dropped at link time. To avoid that, GetCrashpadInfo() has a
reference *back* to that section, which in turn causes the linker to
include it, allowing the note reader to find it at runtime.

The upside of this, is that it's self-contained in the Crashpad code,
and doesn't rely on the user of Crashpad to configure link flags to
ensure that g_crashpad_info appears in the dynamic symbol table of the
final binary/binaries.

Bug: crashpad:196
Change-Id: I599db5903bc98303130d11ad850ba9ceed3b801a