Update patch set 98
Patch Set 98:

Hi all, do you think it's necessary to break this CL up into smaller component CLs in order to be able to review it properly?

I'd be happy to help with peeling off some of the more mechanical reorganization parts if that would be useful.

For the main functionality, maybe we could help this CL by #if'ing some things into crash_report_database.h to allow the Linux/Android/Fuchsia implementation to work and get that part landed, and then do followups to have the Win/Mac/LAF implementations share the common pieces?

Patch-set: 98