fuchsia: Add flock() stub to get test binaries linking again
Because the SDK isn't pinned, this broke over the weekend, though not at
compile time, only at runtime because the symbol disappeared from
libc.so.

Temporarily add a stub local definition here, as excluding
LoggingLockFile() is needed by Settings, which is used downstream by a
bunch of things, and so requires a lot of build file gymnastics to
exclude. Once there's a Fuchsia implementation of Settings, this can be
deleted and LoggingLock/UnlockFile can be #if'd out for Fuchsia.

Bug: crashpad:196
Change-Id: I0971736572b940c8bc2364c01dafc6844f9303d7