Update patch set 1
Patch Set 1: Code-Review+1

LGTM. Is there a Fuchsia-side bug that we can also refer to? That’ll be the best thing to watch for an indication when this is fixed (or of when it might get fixed).

Patch-set: 1
Reviewer: Mark Mentovai <1174099@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=+1