Update patch set 3
Patch Set 3:

> Patch Set 2:
> 
> > Patch Set 2:
> > 
> > (it’d give you another chance to try for a killer hash)
> 
> Haha, I'll take a look at avoiding abspath(). Orobably just need to join() a few relative paths to CRASHPAD_DIR early on.

No luck on the hash.

The abspath()s were mostly because when calling GN we were setting cwd=CRASHPAD_DIR. I thought that was necessary because otherwise it complained that it couldn't find //.gn. But that's overrideable with --root=CRASHPAD_DIR, so switched to that.

Patch-set: 3
Reviewer: Scott Graham <1144260@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>