fuchsia: Fix ninja auto-regen after run
After the avoidance of abspath(), automatic regeneration of ninja files
was broken following a test run. The problem is that the
--runtime-deps-list-file argument gets saved into the regeneration rule,
but it's relative to the cwd. The cwd is CRASHPAD_DIR on the first run,
but the binary_dir on regenerations, so either way it doesn't work (this
should probably fixed in either GN or ninja).

However, it's a bit cluttery to have that --runtime-deps-list-file in
the regeneration rule anyway, when really it's only required to extract
runtime deps at test-running time. So the best thing to do is just
remove it from the regeneration rule by re-running gn gen without the
flag after we've extracted the .runtime_deps to prepare for the run.

Bug: crashpad:196
Change-Id: I009851d8b821fef5c953d463ba9c4880e5cc082a