Update patch set 3
Patch Set 3:

This is an interesting one. The analogue for a normal std::string would be

  std::string s("123");
  assert(&s[3] == '\0');

This is valid for the non-mutating (const) std::basic_string<>::operator[]() where !s.empty(). But it’s not valid for std::basic_string_view::operator[](), and it shouldn’t be valid for base::StringPiece::operator[]() either: in those cases, nobody’s promised that there’s any extra slop following the buffer for NUL-termination or anything else, and in fact the buffer may not be followed by a NUL terminator or any mapped memory at all.

Patch-set: 3
Reviewer: Mark Mentovai <1174099@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>