Update patch set 2
Patch Set 2:

This is surprising because it’d mean a break in binary compatibility. Was this with the exact same sysroot and headers being picked up?

Since this is weird, we should write down the reason for the two symbols and the configurations that result in them. That’s helpful for the future, to prevent someone from removing something that doesn’t appear to be used. Or to frame it more positively: to allow someone to remove something that might in the future become obsolete.

Patch-set: 2
Reviewer: Mark Mentovai <1174099@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>