Update patch set 5
Patch Set 5: Code-Review+1

LGTM, but:

It seemed to me that something (garbage?) was being written to the socket before, because the DCHECK that was being tripped was looking to receive a known message type. That was failing, rather than experiencing an indefinite wait or EOF.

I had to leave before I could finish the investigation. Was that really garbage, or was it something else that was expected to ride along on the same socket after this message?

Patch-set: 5
Reviewer: Mark Mentovai <1174099@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=+1