fuchsia: Implement CaptureContext() for x64
The function prolog is larger on Fuchsia, so the delta to the start of
the function is larger. For reference:

Linux: https://gist.github.com/sgraham/70b202fc7eecf439534c17cf288f5e65
Fuchsia:
https://gist.github.com/sgraham/8c561883508ca7419f78fdf535a7a923

The RSP test only passes when safe-stack is disabled
https://chromium-review.googlesource.com/c/crashpad/crashpad/+/1000140.
Need to investigate if there's a reasonable way to do the sanity test
when safe-stack is on.

Bug: crashpad:196
Change-Id: I2adbcee93c90dbc415309b79e3d16e9c4635f989