fuchsia: Implement CaptureContext() for x64
The function prolog is larger on Fuchsia, so the delta to the start of
the function is larger. For reference:

Linux: https://gist.github.com/sgraham/70b202fc7eecf439534c17cf288f5e65
Fuchsia:
https://gist.github.com/sgraham/8c561883508ca7419f78fdf535a7a923

Fuchsia enables safe-stack by default in the compiler. Disable it for
the test function so that a candidate RSP value can be found by using
the value of locals on the stack.

Bug: crashpad:196
Change-Id: I2adbcee93c90dbc415309b79e3d16e9c4635f989