Reimplement http_transport_test_server in Go
This is a port of http_transport_test_server.py to Go. The motivation
for this is that there is no available Python interpreter on Android or
Fuchsia targets.

In order to accomplish this, some minimal Go toolchain support has been
added to the build for both GN and GYP.

This is not yet enabled on Fuchsia or Android, but neither was the
Python version and this is a step towards enabling it there.

Bug: crashpad:30, crashpad:196, crashpad:227
Change-Id: I2ec92b8c555f4145ae9079c1151a423841ec2812