Implementation of HTTPTransport via raw socket
Currently only handles http, and only POST. Enabled for for Linux and
Fuchsia. Removes libcurl HTTPTransport, since it wasn't available in the
Chromium sysroot anyway.

This is a stepping stone until BoringSSL is available in the Fuchsia
SDK. Once that's available, it should be "relatively straightfoward" to
make http_transport_socket.cc secure its socket using BoringSSL or
OpenSSL depending on the platform.

Bug: crashpad:196, crashpad:227, crashpad:30
Change-Id: If33a0d3f11b9000cbc3f52f96cd024ef274a922f