Update patch set 3
Patch Set 3:

> Note that this is causing bad crash dumps (with *every* memory segment containing 0xfe) being uploaded to crash. See crbug.com/860634 for details.

I just mentioned on bug chromium:860634:

This is what bug crashpad:26 is supposed to address.

That bug would add errors produced during minidump creation to the minidump itself. We’d need to do a little bit of work on the crash processor to extract those messages, and then surface them in a big-enough looks-important sort of way on the crash UI.

I think that’s the cleanest way to handle it, certainly much cleaner and more visible than writing magic bytes or magic words to memory contents that a human may never actually inspect manually.

Patch-set: 3
CC: Gerrit User 1174099 <1174099@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>