Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> This seems fine, but…
> 
> I’m surprised that this is building in Chrome OS at all. We don’t support that platform yet.
> 
> Can you explain a little more about what’s going on in https://crbug.com/873168 that’s causing Crashpad to show up there when it isn’t part of any standard Chrome OS build?

Of course, crashpad is not a dependency for ChromeOS at least not directly, but when we build a ChromeOS image (which is flashed to the device), Chromium is also built together with crashpad. Since our sysroot contains a ptrace header which comes from glibc 2.23, causes a redefinition error and breaks the Chromium build.

Patch-set: 2
Reviewer: Gerrit User 1283828 <1283828@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>