arm/linux: Fix build after a5214472
The build broke because x30 was used instead of r14 in place of LR
which gcc doesn't recognize when building for 64-bit ARM. gcc does
recognize LR for 32-bit ARM, however, so revert to that since it's
more readable.

Also, de-duplicate saving of FP/IP which are synonyms of r11/r12,
saved above.

Change-Id: I8ae28f430cc3c47f4e4cf3679383ed5b94fadd2e