Implement support for PPC64 on Linux
This patch implements support for the PPC64 architecture on Linux hosts.

This is part of the effort to upstream a set of patches that allow
chromium to run on PPC64(LE) hosts available here:
https://wiki.raptorcs.com/wiki/Porting/Chromium.

Note that third_party/mini_chromium needs to be updated with chromium's
architecture definitions for ppc64 in order to build as a standalone
project.

Change-Id: I04e5d88f2619e0a13da1fb2721417b21e0be3693