Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> I think it's more confusing to remove this than to keep it. Do you need to get rid of it for some reason?

I just got curious of the nullptr on the Fuchsia side and then I saw that we had an extra argument for Fuchsia (process_attachments) so I thought we might as well not have this argument we don't use for Fuchsia. But that's it!

Patch-set: 2
Reviewer: Gerrit User 1306103 <1306103@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>