Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> > Patch Set 2:
> > 
> > I think it's more confusing to remove this than to keep it. Do you need to get rid of it for some reason?
> 
> I just got curious of the nullptr on the Fuchsia side and then I saw that we had an extra argument for Fuchsia (process_attachments) so I thought we might as well not have this argument we don't use for Fuchsia. But that's it!

OK, I think it's better to keep it in that case to keep the platforms more similar.

There's still a config where Crashpad builds *for* Fuchsia but not *in* Fuchsia (e.g. Chromium might ship Crashpad as a helper user-space app) where that could be used in theory. (I think it's not too likely that we'll end up shipping that config, but I think for now we should keep the option to do that.)

Patch-set: 2
Reviewer: Gerrit User 1144260 <1144260@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>