Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> It seems like an analogous addition could be useful on other platforms too.

I agree. This change is not backward compatible though. As we control where the Fuchsia code is used, it's fine because I will modify the call site in the rollup CL in the fuchsia tree. For the other platforms, how do we know which are the clients? Additionally we could have overloads I guess.

Patch-set: 1
Reviewer: Gerrit User 1306103 <1306103@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>