Update patch set 1
Patch Set 1:

We’re not terribly concerned with breaking interfaces at this point. Chrome is the only client that we do care about, and we can deal with adding a new parameter when we roll a new version of Crashpad into Chrome.

I don’t think we need an overload. Just an out-param that accepts “nullptr” if the caller doesn’t care should be fine.

Patch-set: 1