Don't auto-generate gyp build files any more
We were still generating gyp build files on the bots, and then using the
GN build. There's some not-working logic in gyp to find the hermetic
Xcode SDK that the new bots are using, so just avoid generating the gyp
build now, since we're not using it anyway.

(My understanding is that the Android build uses
build/gyp_crashpad_android.py instead not build/gyp_crashpad.py. But
there's no Android bots in any case, so it shouldn't be a problem for
that.)

Bug: chromium:892712
Change-Id: Id290f0ddb8d9067deae10b32ab4a8f08a3954ed0
Reviewed-on: https://chromium-review.googlesource.com/c/1287234
Reviewed-by: Mark Mentovai <mark@chromium.org>
Commit-Queue: Scott Graham <scottmg@chromium.org>