Update patch set 3
Patch Set 3:

(2 comments)

> Patch Set 2:
> 
> is there something to do in https://chromium.googlesource.com/crashpad/crashpad/+/master/util/misc/capture_context_fuchsia.S#119? I am not sure the purpose of this file.

Those .S files are used to implement capture of a context without suspending or being in an exception. This is most commonly useful for CrashpadClient::DumpWithoutCrash(). Looking now, I'm not sure that's implemented yet on Fuchsia, so updating the .S could be done separately with a change to make that work.

Patch-set: 3
Reviewer: Gerrit User 1144260 <1144260@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>