Update patch set 9
Patch Set 9:

The Fuchsia LTO build is failing on arm64 with

ld.lld: error: ../../third_party/crashpad/client/crashpad_info.cc:103: improper alignment for relocation R_AARCH64_LDST64_ABS_LO12_NC: 0x5974 is not aligned to 8 bytes
...

I tried upping the alignment in https://chromium-review.googlesource.com/c/crashpad/crashpad/+/1316371 but I think it's not that simple. I guess all those relocs have to be 8-aligned on arm64, but that mucks up the note size so the NoteReader gets confused.

Because of the other changes (pstate/spsr) it's painful to roll Crashpad back in Fuchsia, so I'm going to revert this CL to be able to roll forward instead, and will help reland this on Monday. Sorry!

Patch-set: 9