Update patch set 3
Patch Set 3:

As discussed offline:

I’d prefer it if the ModuleSnapshot::CustomMinidumpStreams() interface were followed, but that doesn’t really apply here because the streams aren’t scoped to individual modules when reading them out of an existing minidump file. (They exist at the module level in ModuleSnapshot because when building a snapshot of a live process, each module is capable of having a copy of Crashpad linked in, and each copy can contribute custom streams.)

That speaks in favor of doing this as ProcessSnapshotMinidump::CustomMinidumpStreams(), following the same interface as used in ModuleSnapshot.

The interface doesn’t _need_ to be CustomMinidumpStreams() as it exists today, but I think that there should be some commonality between the very similar methods on very similar objects. If one changes, the other also ought to.

Patch-set: 3
Reviewer: Gerrit User 1174099 <1174099@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>