Add ProcessMemory::ReadObject().
This function encourages a safer version of creating a copy of a C++ object
across an address space boundary.

In most such usages, the only sane usage is having a
trivially-copyable object. Any other usage will run-afoul of the C++
spec (even if it may work most of the time in practice.  Also, the remote
process data is untrusted and should be validated before use.

This helper function uses a static_assert and an API assumption to
encourage a developer to adopt both constraints when trying to read an
object out of another address space.

Bug: crashpad:
Change-Id: Iff5023677ed7a1aed93ec44ae52cb2a9338cce94